Refactor existing code related to API/AUTH
- splited one firebase file to 3 files
- improve some names of variables
- added enviorment variables for apikeys and ids
