+++ src/redux/notes.js
@@ -1,16 +1,11 @@
-import { db } from "../firebase";
+import { addNoteToDB, addTagToDB } from "../firebase/firebaseAPI";
 
-const ADD_NOTE = "ADD_NOTE";
 const DELETE_NOTE = "DELETE_NOTE";
 const GET_NOTES = "GET_NOTES";
 const GET_TAGS = "GET_TAGS";
 
 export const addNote = note => (dispatch, getState) => {
-  db.collection("notes")
-    .add(note)
-    .then(() => {
-      console.warn("added note");
-    });
+  addNoteToDB(note);
 };
 export const getNotes = notes => {
   return {
@@ -18,12 +13,8 @@ export const getNotes = notes => {
     payload: notes
   };
 };
-export const addTagToDB = tag => (dispatch, getState) => {
-  db.collection("tags")
-    .add(tag)
-    .then(() => {
-      console.warn("added tag");
-    });
+export const addTag = tag => (dispatch, getState) => {
+  addTagToDB(tag);
 };
 export const getTags = tags => {
   return {
@@ -39,11 +30,6 @@ const initialState = {
 
 export const notes = (state = initialState, action) => {
   switch (action.type) {
-    case "ADD_NOTE":
-      return {
-        ...state,
-        notes: [...state.notes, action.payload]
-      };
     case "GET_NOTES":
       return {
         ...state,
